+++ CHANGELOG.md
@@ -5,4 +5,7 @@ History
 -------------------------
 
 * [#2](https://github.com/keyko-io/nevermind-sdk-java/issues/2) Improve the order flow to make use of direct calls instead of flowables
- 
+* [#4](https://github.com/keyko-io/nevermind-sdk-java/issues/4) Changing did prefix to `did:nv`
+* [#5](https://github.com/keyko-io/nevermind-sdk-java/issues/5) Integration tests are not using examples from the Spec repo
+
+